home-mixer: re-anchor entrypoints and portable contracts to upstream e414c17
Restore upstream service composition (HomeMixerConfig -> HomeMixerServer::
build/register), QueryBuilder request ownership, canonical models::{query,
candidate,candidate_features} paths, and upstream component file boundaries
for query hydrators, sources, hydrators, filters, scorers, selectors, and
side effects (HM-E1..E6).

Candidate Pipeline portable contracts follow upstream execution semantics:
synchronous Filter -> FilterResult, per-candidate Vec<Result<C, String>>
hydration/scoring with cardinality protection, run wrappers, and local
logging replacing private instrumentation (U1).

Public proto grows additively only: ForYouFeedQuery/GetForYouFeedV2 wrapper,
DebugScoredPosts with typed stage counts, and RankedFollowing served type.
Debug RPC stays default-off behind HOME_MIXER_ENABLE_DEBUG_RPC plus token.

+++ home-mixer/util/candidates_util.rs
@@ -2,7 +2,7 @@
 //
 // 提供与候选帖子相关的通用辅助函数。
 
-use crate::candidate_pipeline::candidate::PostCandidate;
+use crate::models::candidate::PostCandidate;
 
 /// 获取帖子及其关联帖子的 ID 列表
 ///
@@ -19,13 +19,13 @@ use crate::candidate_pipeline::candidate::PostCandidate;
 ///
 /// # Returns
 /// 帖子本身及其关联帖子的 ID 列表
-pub fn get_related_post_ids(candidate: &PostCandidate) -> Vec<i64> {
+pub fn get_related_post_ids(candidate: &PostCandidate) -> Vec<u64> {
     let mut ids = vec![candidate.tweet_id];
     if let Some(retweeted_id) = candidate.retweeted_tweet_id {
-        ids.push(retweeted_id as i64);
+        ids.push(retweeted_id);
     }
     if let Some(reply_to_id) = candidate.in_reply_to_tweet_id {
-        ids.push(reply_to_id as i64);
+        ids.push(reply_to_id);
     }
     ids
 }
